+++ packages/hippo_utils/lib/src/data_subject/data_subject.dart
@@ -8,11 +8,12 @@
 // ---------------------------------------------------------------------------
 */
 import 'package:flutter/material.dart';
-import 'package:hippo_core_flutter/hippo_core_flutter.dart';
+import 'package:hippo_core/hippo_core.dart';
 
 class SubjectTextField extends StatelessWidget {
   const SubjectTextField({
     super.key,
+    required this.controller,
     required this.subject,
     this.inputDecoration,
     this.keyboardType,
@@ -23,7 +24,8 @@ class SubjectTextField extends StatelessWidget {
     this.maxLines = 1,
   });
 
-  final TextEditingDataSubject subject;
+  final TextEditingController controller;
+  final DataSubject<String> subject;
   final InputDecoration? inputDecoration;
   final TextInputType? keyboardType;
   final bool autocorrect;
@@ -35,7 +37,7 @@ class SubjectTextField extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return TextField(
-      controller: subject.textEditingController,
+      controller: controller,
       decoration: inputDecoration,
       keyboardType: keyboardType,
       autocorrect: autocorrect,